update crawling method
now fetch.rb takes two arguments.

rails runner fetch.rb ARGV[0] ARGV[1]

first argument is a file discriminator.
if you enroll multiple fetch runner instances, this discriminator help you to avoid a xml file collision.
I set numbers (0, 1, 2, 3) for this.

second argument is a course discriminator.
if this is 'U', the fetch instance get undergraduate courses.
if this is 'P', the fetch instance get postgraduate courses.
for this discriminator, the download time of xml file become half.

be careful with 중앙전산원 because of too many requests.